refactor canvasStates to sceneSettings in redux
