+++ baby-gru/src/components/menu-item/MoorhenImportLigandDictionary.tsx
@@ -34,7 +34,7 @@ const MoorhenImportLigandDictionary = (props: {
 
     const dispatch = useDispatch()
     const defaultBondSmoothness = useSelector((state: moorhen.State) => state.sceneSettings.defaultBondSmoothness)
-    const backgroundColor = useSelector((state: moorhen.State) => state.canvasStates.backgroundColor)
+    const backgroundColor = useSelector((state: moorhen.State) => state.sceneSettings.backgroundColor)
     const molecules = useSelector((state: moorhen.State) => state.molecules)
 
     const {
