+++ baby-gru/src/components/menu-item/MoorhenLoadTutorialDataMenuItem.tsx
@@ -20,7 +20,7 @@ export const MoorhenLoadTutorialDataMenuItem = (props: {
 
     const dispatch = useDispatch()
     const defaultBondSmoothness = useSelector((state: moorhen.State) => state.sceneSettings.defaultBondSmoothness)
-    const backgroundColor = useSelector((state: moorhen.State) => state.canvasStates.backgroundColor)
+    const backgroundColor = useSelector((state: moorhen.State) => state.sceneSettings.backgroundColor)
 
     const tutorialNumberSelectorRef = useRef<HTMLSelectElement | null>(null);
