+++ baby-gru/src/components/validation-tools/MoorhenValidation.tsx
@@ -37,9 +37,9 @@ const metricInfoScaling = {
 
 export const MoorhenValidation = (props: Props) => {
     const chartRef = useRef(null);
-    const isDark = useSelector((state: moorhen.State) => state.canvasStates.isDark)
+    const isDark = useSelector((state: moorhen.State) => state.sceneSettings.isDark)
     const molecules = useSelector((state: moorhen.State) => state.molecules)
-    const height = useSelector((state: moorhen.State) => state.canvasStates.height)
+    const height = useSelector((state: moorhen.State) => state.sceneSettings.height)
     const dispatch = useDispatch()
 
     const plugin = {
